requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
psycopg2-binary==2.9.7
Werkzeug==2.3.7
python-dotenv==1.0.0
flask-wtf==1.1.1
PyJWT==2.8.0
flask-migrate==4.0.4
plotly==5.18.0
flask-bcrypt==1.0.1
